Preload images directory
We are now preloading the images directory and storing the
imagemagick objects in a global array.  This helps significantly
reduce the time it takes to handle new requests at the expense
of a slight increase in startup time.  Memory usage will now
be significantly increased dependent on the number of images
preloaded.